bump terrajet

+++ apis/dns/v1alpha1/zz_recordset_types.go
@@ -41,7 +41,7 @@ type RecordsetParameters struct {
 
 	// +kubebuilder:validation:Required
 	// (Optional) The time-to-live of this record set (seconds).
-	TTL *int64 `json:"ttl" tf:"ttl,omitempty"`
+	TTL *float64 `json:"ttl" tf:"ttl,omitempty"`
 
 	// +kubebuilder:validation:Required
 	// (Required) The DNS record set type.
@@ -79,7 +79,7 @@ type RecordsetStatus struct {
 // +kubebuilder:printcolumn:name="EXTERNAL-NAME",type="string",JSONPath=".metadata.annotations.crossplane\\.io/external-name"
 // +kubebuilder:printcolumn:name="AGE",type="date",JSONPath=".metadata.creationTimestamp"
 // +kubebuilder:subresource:status
-// +kubebuilder:resource:scope=Cluster,categories={crossplane,managed,yandex-cloudjet}
+// +kubebuilder:resource:scope=Cluster,categories={crossplane,managed,yandex-cloud}
 type Recordset struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
